refactor(ui-react): re-depth visualization imports after dissolving Components/
Step 3 of the DLS-1022 fold. Dissolving the inner Components/ directory moved
every chart folder up one level, so the 78 specifiers reaching the shared
config.ts and utils/ each lose one `../`:

  '../../config'      -> '../config'        (24)
  '../../../config'   -> '../../config'     (13)
  '../../utils/...'   -> '../utils/...'     (28)
  '../../../utils/...'-> '../../utils/...'  (13)

The rewrite is resolution-aware rather than textual, which matters: every chart
folder has its own component-local utils.ts, so a blind sed on `../utils` would
silently repoint those at the shared module instead of leaving them alone. Each
candidate is only rewritten when the stripped form resolves to a real file
inside visualization/; anything that already resolves is left untouched.

78 specifiers across 51 files, zero unresolved.
