+++ libs/ui-react/src/lib/Components/visualization/Axis/YAxis/YAxis.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 
-import { chartConfig } from '../../../config';
-import { buildTicksData } from '../../../utils/ticks/ticks';
+import { chartConfig } from '../../config';
+import { buildTicksData } from '../../utils/ticks/ticks';
 import { useCartesianChartContext } from '../../CartesianChart/context';
 
 import type { YAxisProps } from './types';
