+++ libs/ui-react/src/lib/Components/visualization/CartesianChart/CartesianChart.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
-import { chartConfig } from '../../config';
+import { chartConfig } from '../config';
 import { MagneticPointsProvider } from '../Point/pointContext';
 import { ScrubberProvider } from '../Scrubber/ScrubberProvider';
 import { CartesianChartProvider, useBuildChartContext } from './context';
