+++ libs/ui-react/src/lib/Components/visualization/DonutChart/hooks/useDonutLoadingAnimation.ts
@@ -1,6 +1,6 @@
 import { useId, useMemo } from 'react';
 
-import { chartConfig } from '../../../config';
+import { chartConfig } from '../../config';
 
 const { loading } = chartConfig.donut;
