+++ libs/ui-react/src/lib/Components/visualization/LineChart/LineChartEmptyState/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { chartConfig } from '../../../config';
+import { chartConfig } from '../../config';
 import { buildPlaceholderTransform } from './utils';
 
 describe('placeholderLinePath', () => {
