+++ libs/ui-react/src/lib/Components/visualization/LineChart/utils.ts
@@ -1,6 +1,6 @@
-import { chartConfig } from '../../config';
-import { isFiniteNumber } from '../../utils/numbers';
-import type { ChartInset, Series } from '../../utils/types';
+import { chartConfig } from '../config';
+import { isFiniteNumber } from '../utils/numbers';
+import type { ChartInset, Series } from '../utils/types';
 import type { XAxisProps } from '../Axis/XAxis';
 import type { YAxisProps } from '../Axis/YAxis';
