+++ libs/ui-react/src/lib/Components/visualization/Point/utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { chartConfig } from '../../config';
+import { chartConfig } from '../config';
 import {
   buildArrowPoints,
   computeLabelX,
