+++ libs/ui-react/src/lib/Components/visualization/Scrubber/types.ts
@@ -1,6 +1,6 @@
 import type { ReactElement, ReactNode, Ref, RefObject, SVGProps } from 'react';
 
-import type { DrawingArea } from '../../utils/types';
+import type { DrawingArea } from '../utils/types';
 import type { ScrubbingOptions } from '../CartesianChart/types';
 
 export type ScrubberContextValue = Required<
