+++ libs/ui-react/src/lib/Components/visualization/Scrubber/useScrubberGeometry.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 
-import { chartConfig } from '../../config';
-import type { DrawingArea } from '../../utils/types';
+import { chartConfig } from '../config';
+import type { DrawingArea } from '../utils/types';
 import { useCartesianChartContext } from '../CartesianChart/context';
 
 import { useScrubberContext } from './context';
